requirement.txt
#python==3.6
tensorflow==1.15
pytorch
onnx
